src/server/requirements.txt
bert-serving-client==1.10.0
Flask==1.1.1
jusText==2.2.0
numpy==1.18.1
scipy==1.4.1
requests==2.22.0
gunicorn
redis
nltk